Fixed SilverStripe detection on non-numeric versions
This broke the Travis builds for forks, since the 'version'
in this case is the branch name, which can't be compared
with version_compare. This lead to the core folder always
being renamed to 'sapphire' (instead of leaving it with
the newer naming of 'framework'). This meant that the test
execution failed because the bootstrap path was unexpected.