Restrict support to actually supported types
This adds a class constant to the BaseInstaller class and all children of the class by which a pattern for matching supported types of packages can be defined. Installer now grabs the value of this constant and uses it in the regular expression check for whether the package type is actually supported after all.