Remove Kirby installers
Kirby 3 doesn't support `composer/installers`, it has its own installer `getkirby/composer-installer`.

Kirby 2 is end-of-life, so no Kirby user is expected to be still using `composer/installers`.

For the way forward it's important that especially the `kirby-plugin` type is removed from `composer/installers` to prevent it from clashing with Kirby's own installer implementation that uses the same package type name.